+++ frontend/src/domains/admin/Login/hooks/useLogin.ts
@@ -1,0 +1,47 @@
+import { postAdminLogin } from '@/apis/admin.api';
+import { ADMIN_BASE, ROUTES } from '@/constants/routes';
+import { useErrorModalContext } from '@/contexts/useErrorModal';
+import { setLocalStorage } from '@/utils/localStorage';
+import { useNavigate } from 'react-router-dom';
+
+interface UseLoginProps {
+  loginValue: {
+    id: string;
+    password: string;
+  };
+}
+
+export default function useLogin({ loginValue }: UseLoginProps) {
+  const navigate = useNavigate();
+  const { showErrorModal } = useErrorModalContext();
+
+  const handleError = (error: Error) => {
+    showErrorModal(
+      error,
+      '로그인에 실패했습니다. 아이디와 비밀번호를 확인해주세요'
+    );
+  };
+
+  const handleSubmit = async (event: React.FormEvent) => {
+    event.preventDefault();
+    try {
+      await postAdminLogin({
+        loginId: loginValue.id,
+        password: loginValue.password,
+        onError: () => {
+          handleError(new Error('로그인 요청 실패'));
+        },
+        onSuccess: (data: Response) => {
+          navigate(ADMIN_BASE + ROUTES.ADMIN_HOME);
+          setLocalStorage('auth', data);
+        },
+      });
+    } catch (error) {
+      handleError(error as Error);
+    }
+  };
+
+  return {
+    handleSubmit,
+  };
+}
